+++ libs/parsing/index/src/lib/auto-complete/completion-for/add-completion-structure-names.ts
@@ -1,3 +1,5 @@
+import { FormatterType, IAssemblerOptions } from '@idl/assembling/config';
+import { AdjustCase, TransformCase } from '@idl/assembling/shared';
 import { CUSTOM_TYPE_DISPLAY_NAMES } from '@idl/data-types/core';
 import { IDL_TRANSLATION } from '@idl/translation';
 import { CompletionItem, CompletionItemKind } from 'vscode-languageserver';
@@ -7,7 +9,10 @@ import { SORT_PRIORITY } from '../sort-priority.interface';
 /**
  * Adds variables to our completion items
  */
-export function AddCompletionStructureNames(complete: CompletionItem[]) {
+export function AddCompletionStructureNames(
+  complete: CompletionItem[],
+  formatting: IAssemblerOptions<FormatterType>
+) {
   // add user procedures first
   const displayNames = Object.values(CUSTOM_TYPE_DISPLAY_NAMES);
   for (let i = 0; i < displayNames.length; i++) {
@@ -15,7 +20,9 @@ export function AddCompletionStructureNames(complete: CompletionItem[]) {
       continue;
     }
     complete.push({
-      label: displayNames[i],
+      label: displayNames[i].startsWith('!')
+        ? AdjustCase(displayNames[i], formatting.style.systemVariables)
+        : TransformCase(displayNames[i], formatting.style.structureNames),
       kind: CompletionItemKind.Class,
       sortText: SORT_PRIORITY.STRUCTURES,
       detail: IDL_TRANSLATION.autoComplete.detail.structure,